requirements.txt
fastapi
uvicorn
kafka-python
redis
psycopg2-binary
sqlalchemy
alembic
python-dotenv
google-generativeai
pydantic
httpx